python/requirements.txt
lightgbm==3.2.1
TA-Lib==0.4.21
requests==2.22.0
pandas==1.3.4
pybitflyer==0.1.9
ccxt==1.84.70
mysql-connector-python==8.0.30
pybit==2.4.0
schedule==1.1.0